src/python-src/requirements.txt
## use the following commands to install requirements
##    pip3 install -r requirements.txt

## for Google Vision Api and Image manipulation
##  install jpeg package: sudo apt-get install libjpeg9-dev
google-api-python-client>=1.5.5
Pillow>=3.4.2

## for text to speech,
gTTS>=1.1.6

## for playing audio sound.
## install simple audio : libasound2-dev 
simpleaudio>=1.0.1

## picamera
picamera>=1.10

## 
RPi.GPIO>=0.6.3
